tidy(ui): organise graph builder files
