+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/addLoRAsToGraph.ts
@@ -1,23 +1,21 @@
 import type { RootState } from 'app/store/store';
 import { zModelIdentifierField } from 'features/nodes/types/common';
-import { filter, size } from 'lodash-es';
-import type { CoreMetadataInvocation, NonNullableGraph, SDXLLoRALoaderInvocation } from 'services/api/types';
-
+import { upsertMetadata } from 'features/nodes/util/graph/canvas/metadata';
 import {
+  CLIP_SKIP,
   LORA_LOADER,
+  MAIN_MODEL_LOADER,
   NEGATIVE_CONDITIONING,
   POSITIVE_CONDITIONING,
-  SDXL_MODEL_LOADER,
-  SDXL_REFINER_INPAINT_CREATE_MASK,
-  SEAMLESS,
-} from './constants';
-import { upsertMetadata } from './metadata';
+} from 'features/nodes/util/graph/constants';
+import { filter, size } from 'lodash-es';
+import type { CoreMetadataInvocation, LoRALoaderInvocation, NonNullableGraph } from 'services/api/types';
 
-export const addSDXLLoRAsToGraph = async (
+export const addLoRAsToGraph = async (
   state: RootState,
   graph: NonNullableGraph,
   baseNodeId: string,
-  modelLoaderNodeId: string = SDXL_MODEL_LOADER
+  modelLoaderNodeId: string = MAIN_MODEL_LOADER
 ): Promise<void> => {
   /**
    * LoRA nodes get the UNet and CLIP models from the main model loader and apply the LoRA to them.
@@ -35,41 +33,36 @@ export const addSDXLLoRAsToGraph = async (
     return;
   }
 
-  const loraMetadata: CoreMetadataInvocation['loras'] = [];
-
-  // Handle Seamless Plugs
-  const unetLoaderId = modelLoaderNodeId;
-  let clipLoaderId = modelLoaderNodeId;
-  if ([SEAMLESS, SDXL_REFINER_INPAINT_CREATE_MASK].includes(modelLoaderNodeId)) {
-    clipLoaderId = SDXL_MODEL_LOADER;
-  }
-
-  // Remove modelLoaderNodeId unet/clip/clip2 connections to feed it to LoRAs
+  // Remove modelLoaderNodeId unet connection to feed it to LoRAs
   graph.edges = graph.edges.filter(
-    (e) =>
-      !(e.source.node_id === unetLoaderId && ['unet'].includes(e.source.field)) &&
-      !(e.source.node_id === clipLoaderId && ['clip'].includes(e.source.field)) &&
-      !(e.source.node_id === clipLoaderId && ['clip2'].includes(e.source.field))
+    (e) => !(e.source.node_id === modelLoaderNodeId && ['unet'].includes(e.source.field))
   );
+  // Remove CLIP_SKIP connections to conditionings to feed it through LoRAs
+  graph.edges = graph.edges.filter((e) => !(e.source.node_id === CLIP_SKIP && ['clip'].includes(e.source.field)));
 
   // we need to remember the last lora so we can chain from it
   let lastLoraNodeId = '';
   let currentLoraIndex = 0;
+  const loraMetadata: CoreMetadataInvocation['loras'] = [];
 
   enabledLoRAs.forEach(async (lora) => {
     const { weight } = lora;
-    const currentLoraNodeId = `${LORA_LOADER}_${lora.model.key}`;
+    const { key } = lora.model;
+    const currentLoraNodeId = `${LORA_LOADER}_${key}`;
     const parsedModel = zModelIdentifierField.parse(lora.model);
 
-    const loraLoaderNode: SDXLLoRALoaderInvocation = {
-      type: 'sdxl_lora_loader',
+    const loraLoaderNode: LoRALoaderInvocation = {
+      type: 'lora_loader',
       id: currentLoraNodeId,
       is_intermediate: true,
       lora: parsedModel,
       weight,
     };
 
-    loraMetadata.push({ model: parsedModel, weight });
+    loraMetadata.push({
+      model: parsedModel,
+      weight,
+    });
 
     // add to graph
     graph.nodes[currentLoraNodeId] = loraLoaderNode;
@@ -77,7 +70,7 @@ export const addSDXLLoRAsToGraph = async (
       // first lora = start the lora chain, attach directly to model loader
       graph.edges.push({
         source: {
-          node_id: unetLoaderId,
+          node_id: modelLoaderNodeId,
           field: 'unet',
         },
         destination: {
@@ -88,23 +81,12 @@ export const addSDXLLoRAsToGraph = async (
 
       graph.edges.push({
         source: {
-          node_id: clipLoaderId,
+          node_id: CLIP_SKIP,
           field: 'clip',
         },
         destination: {
           node_id: currentLoraNodeId,
           field: 'clip',
-        },
-      });
-
-      graph.edges.push({
-        source: {
-          node_id: clipLoaderId,
-          field: 'clip2',
-        },
-        destination: {
-          node_id: currentLoraNodeId,
-          field: 'clip2',
         },
       });
     } else {
@@ -127,17 +109,6 @@ export const addSDXLLoRAsToGraph = async (
         destination: {
           node_id: currentLoraNodeId,
           field: 'clip',
-        },
-      });
-
-      graph.edges.push({
-        source: {
-          node_id: lastLoraNodeId,
-          field: 'clip2',
-        },
-        destination: {
-          node_id: currentLoraNodeId,
-          field: 'clip2',
         },
       });
     }
@@ -174,28 +145,6 @@ export const addSDXLLoRAsToGraph = async (
         destination: {
           node_id: NEGATIVE_CONDITIONING,
           field: 'clip',
-        },
-      });
-
-      graph.edges.push({
-        source: {
-          node_id: currentLoraNodeId,
-          field: 'clip2',
-        },
-        destination: {
-          node_id: POSITIVE_CONDITIONING,
-          field: 'clip2',
-        },
-      });
-
-      graph.edges.push({
-        source: {
-          node_id: currentLoraNodeId,
-          field: 'clip2',
-        },
-        destination: {
-          node_id: NEGATIVE_CONDITIONING,
-          field: 'clip2',
         },
       });
     }
