+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/addSDXLLoRAstoGraph.ts
@@ -1,16 +1,22 @@
 import type { RootState } from 'app/store/store';
 import { zModelIdentifierField } from 'features/nodes/types/common';
+import { upsertMetadata } from 'features/nodes/util/graph/canvas/metadata';
+import {
+  LORA_LOADER,
+  NEGATIVE_CONDITIONING,
+  POSITIVE_CONDITIONING,
+  SDXL_MODEL_LOADER,
+  SDXL_REFINER_INPAINT_CREATE_MASK,
+  SEAMLESS,
+} from 'features/nodes/util/graph/constants';
 import { filter, size } from 'lodash-es';
-import type { CoreMetadataInvocation, LoRALoaderInvocation, NonNullableGraph } from 'services/api/types';
+import type { CoreMetadataInvocation, NonNullableGraph, SDXLLoRALoaderInvocation } from 'services/api/types';
 
-import { CLIP_SKIP, LORA_LOADER, MAIN_MODEL_LOADER, NEGATIVE_CONDITIONING, POSITIVE_CONDITIONING } from './constants';
-import { upsertMetadata } from './metadata';
-
-export const addLoRAsToGraph = async (
+export const addSDXLLoRAsToGraph = async (
   state: RootState,
   graph: NonNullableGraph,
   baseNodeId: string,
-  modelLoaderNodeId: string = MAIN_MODEL_LOADER
+  modelLoaderNodeId: string = SDXL_MODEL_LOADER
 ): Promise<void> => {
   /**
    * LoRA nodes get the UNet and CLIP models from the main model loader and apply the LoRA to them.
@@ -28,36 +34,41 @@ export const addLoRAsToGraph = async (
     return;
   }
 
-  // Remove modelLoaderNodeId unet connection to feed it to LoRAs
+  const loraMetadata: CoreMetadataInvocation['loras'] = [];
+
+  // Handle Seamless Plugs
+  const unetLoaderId = modelLoaderNodeId;
+  let clipLoaderId = modelLoaderNodeId;
+  if ([SEAMLESS, SDXL_REFINER_INPAINT_CREATE_MASK].includes(modelLoaderNodeId)) {
+    clipLoaderId = SDXL_MODEL_LOADER;
+  }
+
+  // Remove modelLoaderNodeId unet/clip/clip2 connections to feed it to LoRAs
   graph.edges = graph.edges.filter(
-    (e) => !(e.source.node_id === modelLoaderNodeId && ['unet'].includes(e.source.field))
+    (e) =>
+      !(e.source.node_id === unetLoaderId && ['unet'].includes(e.source.field)) &&
+      !(e.source.node_id === clipLoaderId && ['clip'].includes(e.source.field)) &&
+      !(e.source.node_id === clipLoaderId && ['clip2'].includes(e.source.field))
   );
-  // Remove CLIP_SKIP connections to conditionings to feed it through LoRAs
-  graph.edges = graph.edges.filter((e) => !(e.source.node_id === CLIP_SKIP && ['clip'].includes(e.source.field)));
 
   // we need to remember the last lora so we can chain from it
   let lastLoraNodeId = '';
   let currentLoraIndex = 0;
-  const loraMetadata: CoreMetadataInvocation['loras'] = [];
 
   enabledLoRAs.forEach(async (lora) => {
     const { weight } = lora;
-    const { key } = lora.model;
-    const currentLoraNodeId = `${LORA_LOADER}_${key}`;
+    const currentLoraNodeId = `${LORA_LOADER}_${lora.model.key}`;
     const parsedModel = zModelIdentifierField.parse(lora.model);
 
-    const loraLoaderNode: LoRALoaderInvocation = {
-      type: 'lora_loader',
+    const loraLoaderNode: SDXLLoRALoaderInvocation = {
+      type: 'sdxl_lora_loader',
       id: currentLoraNodeId,
       is_intermediate: true,
       lora: parsedModel,
       weight,
     };
 
-    loraMetadata.push({
-      model: parsedModel,
-      weight,
-    });
+    loraMetadata.push({ model: parsedModel, weight });
 
     // add to graph
     graph.nodes[currentLoraNodeId] = loraLoaderNode;
@@ -65,7 +76,7 @@ export const addLoRAsToGraph = async (
       // first lora = start the lora chain, attach directly to model loader
       graph.edges.push({
         source: {
-          node_id: modelLoaderNodeId,
+          node_id: unetLoaderId,
           field: 'unet',
         },
         destination: {
@@ -76,12 +87,23 @@ export const addLoRAsToGraph = async (
 
       graph.edges.push({
         source: {
-          node_id: CLIP_SKIP,
+          node_id: clipLoaderId,
           field: 'clip',
         },
         destination: {
           node_id: currentLoraNodeId,
           field: 'clip',
+        },
+      });
+
+      graph.edges.push({
+        source: {
+          node_id: clipLoaderId,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: currentLoraNodeId,
+          field: 'clip2',
         },
       });
     } else {
@@ -104,6 +126,17 @@ export const addLoRAsToGraph = async (
         destination: {
           node_id: currentLoraNodeId,
           field: 'clip',
+        },
+      });
+
+      graph.edges.push({
+        source: {
+          node_id: lastLoraNodeId,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: currentLoraNodeId,
+          field: 'clip2',
         },
       });
     }
@@ -140,6 +173,28 @@ export const addLoRAsToGraph = async (
         destination: {
           node_id: NEGATIVE_CONDITIONING,
           field: 'clip',
+        },
+      });
+
+      graph.edges.push({
+        source: {
+          node_id: currentLoraNodeId,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: POSITIVE_CONDITIONING,
+          field: 'clip2',
+        },
+      });
+
+      graph.edges.push({
+        source: {
+          node_id: currentLoraNodeId,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: NEGATIVE_CONDITIONING,
+          field: 'clip2',
         },
       });
     }
