+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/addWatermarkerToGraph.ts
@@ -1,13 +1,12 @@
 import type { RootState } from 'app/store/store';
+import { LATENTS_TO_IMAGE, NSFW_CHECKER, WATERMARKER } from 'features/nodes/util/graph/constants';
+import { getBoardField, getIsIntermediate } from 'features/nodes/util/graph/graphBuilderUtils';
 import type {
   ImageNSFWBlurInvocation,
   ImageWatermarkInvocation,
   LatentsToImageInvocation,
   NonNullableGraph,
 } from 'services/api/types';
-
-import { LATENTS_TO_IMAGE, NSFW_CHECKER, WATERMARKER } from './constants';
-import { getBoardField, getIsIntermediate } from './graphBuilderUtils';
 
 export const addWatermarkerToGraph = (
   state: RootState,
