+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/buildCanvasGraph.ts
