+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/buildCanvasSDXLInpaintGraph.ts
@@ -1,10 +1,29 @@
 import { logger } from 'app/logging/logger';
 import type { RootState } from 'app/store/store';
+import {
+  CANVAS_OUTPUT,
+  INPAINT_CREATE_MASK,
+  INPAINT_IMAGE,
+  INPAINT_IMAGE_RESIZE_DOWN,
+  INPAINT_IMAGE_RESIZE_UP,
+  LATENTS_TO_IMAGE,
+  MASK_RESIZE_DOWN,
+  MASK_RESIZE_UP,
+  NEGATIVE_CONDITIONING,
+  NOISE,
+  POSITIVE_CONDITIONING,
+  SDXL_CANVAS_INPAINT_GRAPH,
+  SDXL_DENOISE_LATENTS,
+  SDXL_MODEL_LOADER,
+  SDXL_REFINER_SEAMLESS,
+  SEAMLESS,
+} from 'features/nodes/util/graph/constants';
+import { getBoardField, getIsIntermediate, getSDXLStylePrompts } from 'features/nodes/util/graph/graphBuilderUtils';
 import type {
+  CanvasPasteBackInvocation,
+  CreateGradientMaskInvocation,
   ImageDTO,
   ImageToLatentsInvocation,
-  InfillPatchMatchInvocation,
-  InfillTileInvocation,
   NoiseInvocation,
   NonNullableGraph,
 } from 'services/api/types';
@@ -18,37 +37,14 @@ import { addSeamlessToLinearGraph } from './addSeamlessToLinearGraph';
 import { addT2IAdaptersToLinearGraph } from './addT2IAdapterToLinearGraph';
 import { addVAEToGraph } from './addVAEToGraph';
 import { addWatermarkerToGraph } from './addWatermarkerToGraph';
-import {
-  CANVAS_OUTPUT,
-  INPAINT_CREATE_MASK,
-  INPAINT_IMAGE,
-  INPAINT_IMAGE_RESIZE_DOWN,
-  INPAINT_IMAGE_RESIZE_UP,
-  INPAINT_INFILL,
-  INPAINT_INFILL_RESIZE_DOWN,
-  LATENTS_TO_IMAGE,
-  MASK_COMBINE,
-  MASK_FROM_ALPHA,
-  MASK_RESIZE_DOWN,
-  MASK_RESIZE_UP,
-  NEGATIVE_CONDITIONING,
-  NOISE,
-  POSITIVE_CONDITIONING,
-  SDXL_CANVAS_OUTPAINT_GRAPH,
-  SDXL_DENOISE_LATENTS,
-  SDXL_MODEL_LOADER,
-  SDXL_REFINER_SEAMLESS,
-  SEAMLESS,
-} from './constants';
-import { getBoardField, getIsIntermediate, getSDXLStylePrompts } from './graphBuilderUtils';
 
 /**
- * Builds the Canvas tab's Outpaint graph.
+ * Builds the Canvas tab's Inpaint graph.
  */
-export const buildCanvasSDXLOutpaintGraph = async (
+export const buildCanvasSDXLInpaintGraph = async (
   state: RootState,
   canvasInitImage: ImageDTO,
-  canvasMaskImage?: ImageDTO
+  canvasMaskImage: ImageDTO
 ): Promise<NonNullableGraph> => {
   const log = logger('nodes');
   const {
@@ -61,14 +57,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
     seed,
     vaePrecision,
     shouldUseCpuNoise,
-    infillTileSize,
-    infillPatchmatchDownscaleSize,
-    infillMethod,
-    // infillMosaicTileWidth,
-    // infillMosaicTileHeight,
-    // infillMosaicMinColor,
-    // infillMosaicMaxColor,
-    infillColorValue,
     seamlessXAxis,
     seamlessYAxis,
     canvasCoherenceMode,
@@ -91,8 +79,9 @@ export const buildCanvasSDXLOutpaintGraph = async (
   // We may need to set the inpaint width and height to scale the image
   const { scaledBoundingBoxDimensions, boundingBoxScaleMethod } = state.canvas;
 
-  const fp32 = vaePrecision === 'fp32';
   const is_intermediate = true;
+  const fp32 = vaePrecision === 'fp32';
+
   const isUsingScaledDimensions = ['auto', 'manual'].includes(boundingBoxScaleMethod);
 
   let modelLoaderNodeId = SDXL_MODEL_LOADER;
@@ -103,11 +92,12 @@ export const buildCanvasSDXLOutpaintGraph = async (
   const { positiveStylePrompt, negativeStylePrompt } = getSDXLStylePrompts(state);
 
   const graph: NonNullableGraph = {
-    id: SDXL_CANVAS_OUTPAINT_GRAPH,
+    id: SDXL_CANVAS_INPAINT_GRAPH,
     nodes: {
-      [SDXL_MODEL_LOADER]: {
+      [modelLoaderNodeId]: {
         type: 'sdxl_model_loader',
-        id: SDXL_MODEL_LOADER,
+        id: modelLoaderNodeId,
+        is_intermediate,
         model,
       },
       [POSITIVE_CONDITIONING]: {
@@ -123,18 +113,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
         is_intermediate,
         prompt: negativePrompt,
         style: negativeStylePrompt,
-      },
-      [MASK_FROM_ALPHA]: {
-        type: 'tomask',
-        id: MASK_FROM_ALPHA,
-        is_intermediate,
-        image: canvasInitImage,
-      },
-      [MASK_COMBINE]: {
-        type: 'mask_combine',
-        id: MASK_COMBINE,
-        is_intermediate,
-        mask2: canvasMaskImage,
       },
       [INPAINT_IMAGE]: {
         type: 'i2l',
@@ -154,8 +132,8 @@ export const buildCanvasSDXLOutpaintGraph = async (
         id: INPAINT_CREATE_MASK,
         is_intermediate,
         coherence_mode: canvasCoherenceMode,
-        edge_radius: canvasCoherenceEdgeSize,
         minimum_denoise: refinerModel ? Math.max(0.2, canvasCoherenceMinDenoise) : canvasCoherenceMinDenoise,
+        edge_radius: canvasCoherenceEdgeSize,
         tiled: false,
         fp32: fp32,
       },
@@ -170,7 +148,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
         denoising_start: refinerModel ? Math.min(refinerStart, 1 - strength) : 1 - strength,
         denoising_end: refinerModel ? refinerStart : 1,
       },
-
       [LATENTS_TO_IMAGE]: {
         type: 'l2i',
         id: LATENTS_TO_IMAGE,
@@ -182,15 +159,15 @@ export const buildCanvasSDXLOutpaintGraph = async (
         id: CANVAS_OUTPUT,
         is_intermediate: getIsIntermediate(state),
         board: getBoardField(state),
-        use_cache: false,
         mask_blur: maskBlur,
+        source_image: canvasInitImage,
       },
     },
     edges: [
-      // Connect Model Loader To UNet and CLIP
+      // Connect Model Loader to UNet and CLIP
       {
         source: {
-          node_id: SDXL_MODEL_LOADER,
+          node_id: modelLoaderNodeId,
           field: 'unet',
         },
         destination: {
@@ -200,7 +177,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
       },
       {
         source: {
-          node_id: SDXL_MODEL_LOADER,
+          node_id: modelLoaderNodeId,
           field: 'clip',
         },
         destination: {
@@ -210,7 +187,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
       },
       {
         source: {
-          node_id: SDXL_MODEL_LOADER,
+          node_id: modelLoaderNodeId,
           field: 'clip2',
         },
         destination: {
@@ -220,7 +197,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
       },
       {
         source: {
-          node_id: SDXL_MODEL_LOADER,
+          node_id: modelLoaderNodeId,
           field: 'clip',
         },
         destination: {
@@ -230,7 +207,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
       },
       {
         source: {
-          node_id: SDXL_MODEL_LOADER,
+          node_id: modelLoaderNodeId,
           field: 'clip2',
         },
         destination: {
@@ -248,29 +225,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
           field: 'unet',
         },
       },
-      // Connect Infill Result To Inpaint Image
-      {
-        source: {
-          node_id: INPAINT_INFILL,
-          field: 'image',
-        },
-        destination: {
-          node_id: INPAINT_IMAGE,
-          field: 'image',
-        },
-      },
-      // Combine Mask from Init Image with User Painted Mask
-      {
-        source: {
-          node_id: MASK_FROM_ALPHA,
-          field: 'image',
-        },
-        destination: {
-          node_id: MASK_COMBINE,
-          field: 'mask1',
-        },
-      },
-      // Plug Everything Into Inpaint Node
+      // Connect Everything To Inpaint Node
       {
         source: {
           node_id: POSITIVE_CONDITIONING,
@@ -311,17 +266,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
           field: 'latents',
         },
       },
-      // Create Inpaint Mask
-      {
-        source: {
-          node_id: isUsingScaledDimensions ? MASK_RESIZE_UP : MASK_COMBINE,
-          field: 'image',
-        },
-        destination: {
-          node_id: INPAINT_CREATE_MASK,
-          field: 'mask',
-        },
-      },
       {
         source: {
           node_id: INPAINT_CREATE_MASK,
@@ -332,7 +276,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
           field: 'denoise_mask',
         },
       },
-
+      // Decode Inpainted Latents To Image
       {
         source: {
           node_id: SDXL_DENOISE_LATENTS,
@@ -345,63 +289,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
       },
     ],
   };
-
-  // Add Infill Nodes
-  if (infillMethod === 'patchmatch') {
-    graph.nodes[INPAINT_INFILL] = {
-      type: 'infill_patchmatch',
-      id: INPAINT_INFILL,
-      is_intermediate,
-      downscale: infillPatchmatchDownscaleSize,
-    };
-  }
-
-  if (infillMethod === 'lama') {
-    graph.nodes[INPAINT_INFILL] = {
-      type: 'infill_lama',
-      id: INPAINT_INFILL,
-      is_intermediate,
-    };
-  }
-
-  if (infillMethod === 'cv2') {
-    graph.nodes[INPAINT_INFILL] = {
-      type: 'infill_cv2',
-      id: INPAINT_INFILL,
-      is_intermediate,
-    };
-  }
-
-  if (infillMethod === 'tile') {
-    graph.nodes[INPAINT_INFILL] = {
-      type: 'infill_tile',
-      id: INPAINT_INFILL,
-      is_intermediate,
-      tile_size: infillTileSize,
-    };
-  }
-
-  // TODO: add mosaic back
-  // if (infillMethod === 'mosaic') {
-  //   graph.nodes[INPAINT_INFILL] = {
-  //     type: 'infill_mosaic',
-  //     id: INPAINT_INFILL,
-  //     is_intermediate,
-  //     tile_width: infillMosaicTileWidth,
-  //     tile_height: infillMosaicTileHeight,
-  //     min_color: infillMosaicMinColor,
-  //     max_color: infillMosaicMaxColor,
-  //   };
-  // }
-
-  if (infillMethod === 'color') {
-    graph.nodes[INPAINT_INFILL] = {
-      type: 'infill_rgba',
-      id: INPAINT_INFILL,
-      is_intermediate,
-      color: infillColorValue,
-    };
-  }
 
   // Handle Scale Before Processing
   if (isUsingScaledDimensions) {
@@ -423,17 +310,11 @@ export const buildCanvasSDXLOutpaintGraph = async (
       is_intermediate,
       width: scaledWidth,
       height: scaledHeight,
+      image: canvasMaskImage,
     };
     graph.nodes[INPAINT_IMAGE_RESIZE_DOWN] = {
       type: 'img_resize',
       id: INPAINT_IMAGE_RESIZE_DOWN,
-      is_intermediate,
-      width: width,
-      height: height,
-    };
-    graph.nodes[INPAINT_INFILL_RESIZE_DOWN] = {
-      type: 'img_resize',
-      id: INPAINT_INFILL_RESIZE_DOWN,
       is_intermediate,
       width: width,
       height: height,
@@ -451,15 +332,25 @@ export const buildCanvasSDXLOutpaintGraph = async (
 
     // Connect Nodes
     graph.edges.push(
-      // Scale Inpaint Image
+      // Scale Inpaint Image and Mask
       {
         source: {
           node_id: INPAINT_IMAGE_RESIZE_UP,
           field: 'image',
         },
         destination: {
-          node_id: INPAINT_INFILL,
+          node_id: INPAINT_IMAGE,
           field: 'image',
+        },
+      },
+      {
+        source: {
+          node_id: MASK_RESIZE_UP,
+          field: 'image',
+        },
+        destination: {
+          node_id: INPAINT_CREATE_MASK,
+          field: 'mask',
         },
       },
       {
@@ -472,18 +363,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
           field: 'image',
         },
       },
-      // Take combined mask and resize
-      {
-        source: {
-          node_id: MASK_COMBINE,
-          field: 'image',
-        },
-        destination: {
-          node_id: MASK_RESIZE_UP,
-          field: 'image',
-        },
-      },
-      // Resize Results Down
+      // Resize Down
       {
         source: {
           node_id: LATENTS_TO_IMAGE,
@@ -504,27 +384,7 @@ export const buildCanvasSDXLOutpaintGraph = async (
           field: 'image',
         },
       },
-      {
-        source: {
-          node_id: INPAINT_INFILL,
-          field: 'image',
-        },
-        destination: {
-          node_id: INPAINT_INFILL_RESIZE_DOWN,
-          field: 'image',
-        },
-      },
       // Paste Back
-      {
-        source: {
-          node_id: INPAINT_INFILL_RESIZE_DOWN,
-          field: 'image',
-        },
-        destination: {
-          node_id: CANVAS_OUTPUT,
-          field: 'source_image',
-        },
-      },
       {
         source: {
           node_id: INPAINT_IMAGE_RESIZE_DOWN,
@@ -548,11 +408,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
     );
   } else {
     // Add Images To Nodes
-    graph.nodes[INPAINT_INFILL] = {
-      ...(graph.nodes[INPAINT_INFILL] as InfillTileInvocation | InfillPatchMatchInvocation),
-      image: canvasInitImage,
-    };
-
     (graph.nodes[NOISE] as NoiseInvocation).width = width;
     (graph.nodes[NOISE] as NoiseInvocation).height = height;
 
@@ -561,38 +416,27 @@ export const buildCanvasSDXLOutpaintGraph = async (
       image: canvasInitImage,
     };
 
-    graph.edges.push(
-      {
-        source: {
-          node_id: INPAINT_INFILL,
-          field: 'image',
-        },
-        destination: {
-          node_id: CANVAS_OUTPUT,
-          field: 'source_image',
-        },
+    graph.nodes[INPAINT_CREATE_MASK] = {
+      ...(graph.nodes[INPAINT_CREATE_MASK] as CreateGradientMaskInvocation),
+      mask: canvasMaskImage,
+    };
+
+    // Paste Back
+    graph.nodes[CANVAS_OUTPUT] = {
+      ...(graph.nodes[CANVAS_OUTPUT] as CanvasPasteBackInvocation),
+      mask: canvasMaskImage,
+    };
+
+    graph.edges.push({
+      source: {
+        node_id: LATENTS_TO_IMAGE,
+        field: 'image',
       },
-      {
-        source: {
-          node_id: LATENTS_TO_IMAGE,
-          field: 'image',
-        },
-        destination: {
-          node_id: CANVAS_OUTPUT,
-          field: 'target_image',
-        },
+      destination: {
+        node_id: CANVAS_OUTPUT,
+        field: 'target_image',
       },
-      {
-        source: {
-          node_id: INPAINT_CREATE_MASK,
-          field: 'expanded_mask_area',
-        },
-        destination: {
-          node_id: CANVAS_OUTPUT,
-          field: 'mask',
-        },
-      }
-    );
+    });
   }
 
   // Add Seamless To Graph
@@ -620,7 +464,6 @@ export const buildCanvasSDXLOutpaintGraph = async (
 
   // Add IP Adapter
   await addIPAdapterToLinearGraph(state, graph, SDXL_DENOISE_LATENTS);
-
   await addT2IAdaptersToLinearGraph(state, graph, SDXL_DENOISE_LATENTS);
 
   // NSFW & watermark - must be last thing added to graph
