+++ invokeai/frontend/web/src/features/nodes/util/graph/canvas/buildCanvasSDXLOutpaintGraph.ts
@@ -1,49 +1,54 @@
 import { logger } from 'app/logging/logger';
 import type { RootState } from 'app/store/store';
+import {
+  CANVAS_OUTPUT,
+  INPAINT_CREATE_MASK,
+  INPAINT_IMAGE,
+  INPAINT_IMAGE_RESIZE_DOWN,
+  INPAINT_IMAGE_RESIZE_UP,
+  INPAINT_INFILL,
+  INPAINT_INFILL_RESIZE_DOWN,
+  LATENTS_TO_IMAGE,
+  MASK_COMBINE,
+  MASK_FROM_ALPHA,
+  MASK_RESIZE_DOWN,
+  MASK_RESIZE_UP,
+  NEGATIVE_CONDITIONING,
+  NOISE,
+  POSITIVE_CONDITIONING,
+  SDXL_CANVAS_OUTPAINT_GRAPH,
+  SDXL_DENOISE_LATENTS,
+  SDXL_MODEL_LOADER,
+  SDXL_REFINER_SEAMLESS,
+  SEAMLESS,
+} from 'features/nodes/util/graph/constants';
+import { getBoardField, getIsIntermediate, getSDXLStylePrompts } from 'features/nodes/util/graph/graphBuilderUtils';
 import type {
-  CanvasPasteBackInvocation,
-  CreateGradientMaskInvocation,
   ImageDTO,
   ImageToLatentsInvocation,
+  InfillPatchMatchInvocation,
+  InfillTileInvocation,
   NoiseInvocation,
   NonNullableGraph,
 } from 'services/api/types';
 
 import { addControlNetToLinearGraph } from './addControlNetToLinearGraph';
 import { addIPAdapterToLinearGraph } from './addIPAdapterToLinearGraph';
-import { addLoRAsToGraph } from './addLoRAsToGraph';
 import { addNSFWCheckerToGraph } from './addNSFWCheckerToGraph';
+import { addSDXLLoRAsToGraph } from './addSDXLLoRAstoGraph';
+import { addSDXLRefinerToGraph } from './addSDXLRefinerToGraph';
 import { addSeamlessToLinearGraph } from './addSeamlessToLinearGraph';
 import { addT2IAdaptersToLinearGraph } from './addT2IAdapterToLinearGraph';
 import { addVAEToGraph } from './addVAEToGraph';
 import { addWatermarkerToGraph } from './addWatermarkerToGraph';
-import {
-  CANVAS_INPAINT_GRAPH,
-  CANVAS_OUTPUT,
-  CLIP_SKIP,
-  DENOISE_LATENTS,
-  INPAINT_CREATE_MASK,
-  INPAINT_IMAGE,
-  INPAINT_IMAGE_RESIZE_DOWN,
-  INPAINT_IMAGE_RESIZE_UP,
-  LATENTS_TO_IMAGE,
-  MAIN_MODEL_LOADER,
-  MASK_RESIZE_DOWN,
-  MASK_RESIZE_UP,
-  NEGATIVE_CONDITIONING,
-  NOISE,
-  POSITIVE_CONDITIONING,
-  SEAMLESS,
-} from './constants';
-import { getBoardField, getIsIntermediate } from './graphBuilderUtils';
 
 /**
- * Builds the Canvas tab's Inpaint graph.
+ * Builds the Canvas tab's Outpaint graph.
  */
-export const buildCanvasInpaintGraph = async (
+export const buildCanvasSDXLOutpaintGraph = async (
   state: RootState,
   canvasInitImage: ImageDTO,
-  canvasMaskImage: ImageDTO
+  canvasMaskImage?: ImageDTO
 ): Promise<NonNullableGraph> => {
   const log = logger('nodes');
   const {
@@ -56,7 +61,14 @@ export const buildCanvasInpaintGraph = async (
     seed,
     vaePrecision,
     shouldUseCpuNoise,
-    clipSkip,
+    infillTileSize,
+    infillPatchmatchDownscaleSize,
+    infillMethod,
+    // infillMosaicTileWidth,
+    // infillMosaicTileHeight,
+    // infillMosaicMinColor,
+    // infillMosaicMaxColor,
+    infillColorValue,
     seamlessXAxis,
     seamlessYAxis,
     canvasCoherenceMode,
@@ -65,6 +77,8 @@ export const buildCanvasInpaintGraph = async (
     maskBlur,
   } = state.generation;
   const { positivePrompt, negativePrompt } = state.controlLayers.present;
+
+  const { refinerModel, refinerStart } = state.sdxl;
 
   if (!model) {
     log.error('No model found in state');
@@ -76,41 +90,51 @@ export const buildCanvasInpaintGraph = async (
 
   // We may need to set the inpaint width and height to scale the image
   const { scaledBoundingBoxDimensions, boundingBoxScaleMethod } = state.canvas;
-  const is_intermediate = true;
-  const fp32 = vaePrecision === 'fp32';
 
+  const fp32 = vaePrecision === 'fp32';
+  const is_intermediate = true;
   const isUsingScaledDimensions = ['auto', 'manual'].includes(boundingBoxScaleMethod);
 
-  let modelLoaderNodeId = MAIN_MODEL_LOADER;
+  let modelLoaderNodeId = SDXL_MODEL_LOADER;
 
   const use_cpu = shouldUseCpuNoise;
 
+  // Construct Style Prompt
+  const { positiveStylePrompt, negativeStylePrompt } = getSDXLStylePrompts(state);
+
   const graph: NonNullableGraph = {
-    id: CANVAS_INPAINT_GRAPH,
+    id: SDXL_CANVAS_OUTPAINT_GRAPH,
     nodes: {
-      [modelLoaderNodeId]: {
-        type: 'main_model_loader',
-        id: modelLoaderNodeId,
-        is_intermediate,
+      [SDXL_MODEL_LOADER]: {
+        type: 'sdxl_model_loader',
+        id: SDXL_MODEL_LOADER,
         model,
       },
-      [CLIP_SKIP]: {
-        type: 'clip_skip',
-        id: CLIP_SKIP,
-        is_intermediate,
-        skipped_layers: clipSkip,
-      },
       [POSITIVE_CONDITIONING]: {
-        type: 'compel',
+        type: 'sdxl_compel_prompt',
         id: POSITIVE_CONDITIONING,
         is_intermediate,
         prompt: positivePrompt,
+        style: positiveStylePrompt,
       },
       [NEGATIVE_CONDITIONING]: {
-        type: 'compel',
+        type: 'sdxl_compel_prompt',
         id: NEGATIVE_CONDITIONING,
         is_intermediate,
         prompt: negativePrompt,
+        style: negativeStylePrompt,
+      },
+      [MASK_FROM_ALPHA]: {
+        type: 'tomask',
+        id: MASK_FROM_ALPHA,
+        is_intermediate,
+        image: canvasInitImage,
+      },
+      [MASK_COMBINE]: {
+        type: 'mask_combine',
+        id: MASK_COMBINE,
+        is_intermediate,
+        mask2: canvasMaskImage,
       },
       [INPAINT_IMAGE]: {
         type: 'i2l',
@@ -130,22 +154,23 @@ export const buildCanvasInpaintGraph = async (
         id: INPAINT_CREATE_MASK,
         is_intermediate,
         coherence_mode: canvasCoherenceMode,
-        minimum_denoise: canvasCoherenceMinDenoise,
         edge_radius: canvasCoherenceEdgeSize,
+        minimum_denoise: refinerModel ? Math.max(0.2, canvasCoherenceMinDenoise) : canvasCoherenceMinDenoise,
         tiled: false,
         fp32: fp32,
       },
-      [DENOISE_LATENTS]: {
+      [SDXL_DENOISE_LATENTS]: {
         type: 'denoise_latents',
-        id: DENOISE_LATENTS,
+        id: SDXL_DENOISE_LATENTS,
         is_intermediate,
         steps: steps,
         cfg_scale: cfg_scale,
         cfg_rescale_multiplier,
         scheduler: scheduler,
-        denoising_start: 1 - strength,
-        denoising_end: 1,
+        denoising_start: refinerModel ? Math.min(refinerStart, 1 - strength) : 1 - strength,
+        denoising_end: refinerModel ? refinerStart : 1,
       },
+
       [LATENTS_TO_IMAGE]: {
         type: 'l2i',
         id: LATENTS_TO_IMAGE,
@@ -157,30 +182,60 @@ export const buildCanvasInpaintGraph = async (
         id: CANVAS_OUTPUT,
         is_intermediate: getIsIntermediate(state),
         board: getBoardField(state),
+        use_cache: false,
         mask_blur: maskBlur,
-        source_image: canvasInitImage,
       },
     },
     edges: [
-      // Connect Model Loader to CLIP Skip and UNet
+      // Connect Model Loader To UNet and CLIP
       {
         source: {
-          node_id: modelLoaderNodeId,
+          node_id: SDXL_MODEL_LOADER,
           field: 'unet',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'unet',
         },
       },
       {
         source: {
-          node_id: modelLoaderNodeId,
+          node_id: SDXL_MODEL_LOADER,
           field: 'clip',
         },
         destination: {
-          node_id: CLIP_SKIP,
+          node_id: POSITIVE_CONDITIONING,
           field: 'clip',
+        },
+      },
+      {
+        source: {
+          node_id: SDXL_MODEL_LOADER,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: POSITIVE_CONDITIONING,
+          field: 'clip2',
+        },
+      },
+      {
+        source: {
+          node_id: SDXL_MODEL_LOADER,
+          field: 'clip',
+        },
+        destination: {
+          node_id: NEGATIVE_CONDITIONING,
+          field: 'clip',
+        },
+      },
+      {
+        source: {
+          node_id: SDXL_MODEL_LOADER,
+          field: 'clip2',
+        },
+        destination: {
+          node_id: NEGATIVE_CONDITIONING,
+          field: 'clip2',
         },
       },
       {
@@ -193,35 +248,36 @@ export const buildCanvasInpaintGraph = async (
           field: 'unet',
         },
       },
-      // Connect CLIP Skip to Conditioning
+      // Connect Infill Result To Inpaint Image
       {
         source: {
-          node_id: CLIP_SKIP,
-          field: 'clip',
+          node_id: INPAINT_INFILL,
+          field: 'image',
         },
         destination: {
-          node_id: POSITIVE_CONDITIONING,
-          field: 'clip',
+          node_id: INPAINT_IMAGE,
+          field: 'image',
         },
       },
+      // Combine Mask from Init Image with User Painted Mask
       {
         source: {
-          node_id: CLIP_SKIP,
-          field: 'clip',
+          node_id: MASK_FROM_ALPHA,
+          field: 'image',
         },
         destination: {
-          node_id: NEGATIVE_CONDITIONING,
-          field: 'clip',
+          node_id: MASK_COMBINE,
+          field: 'mask1',
         },
       },
-      // Connect Everything To Inpaint Node
+      // Plug Everything Into Inpaint Node
       {
         source: {
           node_id: POSITIVE_CONDITIONING,
           field: 'conditioning',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'positive_conditioning',
         },
       },
@@ -231,7 +287,7 @@ export const buildCanvasInpaintGraph = async (
           field: 'conditioning',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'negative_conditioning',
         },
       },
@@ -241,7 +297,7 @@ export const buildCanvasInpaintGraph = async (
           field: 'noise',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'noise',
         },
       },
@@ -251,8 +307,19 @@ export const buildCanvasInpaintGraph = async (
           field: 'latents',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'latents',
+        },
+      },
+      // Create Inpaint Mask
+      {
+        source: {
+          node_id: isUsingScaledDimensions ? MASK_RESIZE_UP : MASK_COMBINE,
+          field: 'image',
+        },
+        destination: {
+          node_id: INPAINT_CREATE_MASK,
+          field: 'mask',
         },
       },
       {
@@ -261,14 +328,14 @@ export const buildCanvasInpaintGraph = async (
           field: 'denoise_mask',
         },
         destination: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'denoise_mask',
         },
       },
-      // Decode Inpainted Latents To Image
+
       {
         source: {
-          node_id: DENOISE_LATENTS,
+          node_id: SDXL_DENOISE_LATENTS,
           field: 'latents',
         },
         destination: {
@@ -278,6 +345,63 @@ export const buildCanvasInpaintGraph = async (
       },
     ],
   };
+
+  // Add Infill Nodes
+  if (infillMethod === 'patchmatch') {
+    graph.nodes[INPAINT_INFILL] = {
+      type: 'infill_patchmatch',
+      id: INPAINT_INFILL,
+      is_intermediate,
+      downscale: infillPatchmatchDownscaleSize,
+    };
+  }
+
+  if (infillMethod === 'lama') {
+    graph.nodes[INPAINT_INFILL] = {
+      type: 'infill_lama',
+      id: INPAINT_INFILL,
+      is_intermediate,
+    };
+  }
+
+  if (infillMethod === 'cv2') {
+    graph.nodes[INPAINT_INFILL] = {
+      type: 'infill_cv2',
+      id: INPAINT_INFILL,
+      is_intermediate,
+    };
+  }
+
+  if (infillMethod === 'tile') {
+    graph.nodes[INPAINT_INFILL] = {
+      type: 'infill_tile',
+      id: INPAINT_INFILL,
+      is_intermediate,
+      tile_size: infillTileSize,
+    };
+  }
+
+  // TODO: add mosaic back
+  // if (infillMethod === 'mosaic') {
+  //   graph.nodes[INPAINT_INFILL] = {
+  //     type: 'infill_mosaic',
+  //     id: INPAINT_INFILL,
+  //     is_intermediate,
+  //     tile_width: infillMosaicTileWidth,
+  //     tile_height: infillMosaicTileHeight,
+  //     min_color: infillMosaicMinColor,
+  //     max_color: infillMosaicMaxColor,
+  //   };
+  // }
+
+  if (infillMethod === 'color') {
+    graph.nodes[INPAINT_INFILL] = {
+      type: 'infill_rgba',
+      id: INPAINT_INFILL,
+      is_intermediate,
+      color: infillColorValue,
+    };
+  }
 
   // Handle Scale Before Processing
   if (isUsingScaledDimensions) {
@@ -299,11 +423,17 @@ export const buildCanvasInpaintGraph = async (
       is_intermediate,
       width: scaledWidth,
       height: scaledHeight,
-      image: canvasMaskImage,
     };
     graph.nodes[INPAINT_IMAGE_RESIZE_DOWN] = {
       type: 'img_resize',
       id: INPAINT_IMAGE_RESIZE_DOWN,
+      is_intermediate,
+      width: width,
+      height: height,
+    };
+    graph.nodes[INPAINT_INFILL_RESIZE_DOWN] = {
+      type: 'img_resize',
+      id: INPAINT_INFILL_RESIZE_DOWN,
       is_intermediate,
       width: width,
       height: height,
@@ -321,38 +451,39 @@ export const buildCanvasInpaintGraph = async (
 
     // Connect Nodes
     graph.edges.push(
-      // Scale Inpaint Image and Mask
+      // Scale Inpaint Image
       {
         source: {
           node_id: INPAINT_IMAGE_RESIZE_UP,
           field: 'image',
         },
         destination: {
-          node_id: INPAINT_IMAGE,
+          node_id: INPAINT_INFILL,
           field: 'image',
         },
       },
       {
         source: {
+          node_id: INPAINT_IMAGE_RESIZE_UP,
+          field: 'image',
+        },
+        destination: {
+          node_id: INPAINT_CREATE_MASK,
+          field: 'image',
+        },
+      },
+      // Take combined mask and resize
+      {
+        source: {
+          node_id: MASK_COMBINE,
+          field: 'image',
+        },
+        destination: {
           node_id: MASK_RESIZE_UP,
           field: 'image',
         },
-        destination: {
-          node_id: INPAINT_CREATE_MASK,
-          field: 'mask',
-        },
       },
-      {
-        source: {
-          node_id: INPAINT_IMAGE_RESIZE_UP,
-          field: 'image',
-        },
-        destination: {
-          node_id: INPAINT_CREATE_MASK,
-          field: 'image',
-        },
-      },
-      // Resize Down
+      // Resize Results Down
       {
         source: {
           node_id: LATENTS_TO_IMAGE,
@@ -373,7 +504,27 @@ export const buildCanvasInpaintGraph = async (
           field: 'image',
         },
       },
+      {
+        source: {
+          node_id: INPAINT_INFILL,
+          field: 'image',
+        },
+        destination: {
+          node_id: INPAINT_INFILL_RESIZE_DOWN,
+          field: 'image',
+        },
+      },
       // Paste Back
+      {
+        source: {
+          node_id: INPAINT_INFILL_RESIZE_DOWN,
+          field: 'image',
+        },
+        destination: {
+          node_id: CANVAS_OUTPUT,
+          field: 'source_image',
+        },
+      },
       {
         source: {
           node_id: INPAINT_IMAGE_RESIZE_DOWN,
@@ -397,6 +548,11 @@ export const buildCanvasInpaintGraph = async (
     );
   } else {
     // Add Images To Nodes
+    graph.nodes[INPAINT_INFILL] = {
+      ...(graph.nodes[INPAINT_INFILL] as InfillTileInvocation | InfillPatchMatchInvocation),
+      image: canvasInitImage,
+    };
+
     (graph.nodes[NOISE] as NoiseInvocation).width = width;
     (graph.nodes[NOISE] as NoiseInvocation).height = height;
 
@@ -405,27 +561,38 @@ export const buildCanvasInpaintGraph = async (
       image: canvasInitImage,
     };
 
-    graph.nodes[INPAINT_CREATE_MASK] = {
-      ...(graph.nodes[INPAINT_CREATE_MASK] as CreateGradientMaskInvocation),
-      mask: canvasMaskImage,
-    };
-
-    // Paste Back
-    graph.nodes[CANVAS_OUTPUT] = {
-      ...(graph.nodes[CANVAS_OUTPUT] as CanvasPasteBackInvocation),
-      mask: canvasMaskImage,
-    };
-
-    graph.edges.push({
-      source: {
-        node_id: LATENTS_TO_IMAGE,
-        field: 'image',
+    graph.edges.push(
+      {
+        source: {
+          node_id: INPAINT_INFILL,
+          field: 'image',
+        },
+        destination: {
+          node_id: CANVAS_OUTPUT,
+          field: 'source_image',
+        },
       },
-      destination: {
-        node_id: CANVAS_OUTPUT,
-        field: 'target_image',
+      {
+        source: {
+          node_id: LATENTS_TO_IMAGE,
+          field: 'image',
+        },
+        destination: {
+          node_id: CANVAS_OUTPUT,
+          field: 'target_image',
+        },
       },
-    });
+      {
+        source: {
+          node_id: INPAINT_CREATE_MASK,
+          field: 'expanded_mask_area',
+        },
+        destination: {
+          node_id: CANVAS_OUTPUT,
+          field: 'mask',
+        },
+      }
+    );
   }
 
   // Add Seamless To Graph
@@ -434,18 +601,28 @@ export const buildCanvasInpaintGraph = async (
     modelLoaderNodeId = SEAMLESS;
   }
 
+  // Add Refiner if enabled
+  if (refinerModel) {
+    await addSDXLRefinerToGraph(state, graph, SDXL_DENOISE_LATENTS, modelLoaderNodeId);
+    if (seamlessXAxis || seamlessYAxis) {
+      modelLoaderNodeId = SDXL_REFINER_SEAMLESS;
+    }
+  }
+
   // Add VAE
   await addVAEToGraph(state, graph, modelLoaderNodeId);
 
   // add LoRA support
-  await addLoRAsToGraph(state, graph, DENOISE_LATENTS, modelLoaderNodeId);
+  await addSDXLLoRAsToGraph(state, graph, SDXL_DENOISE_LATENTS, modelLoaderNodeId);
 
   // add controlnet, mutating `graph`
-  await addControlNetToLinearGraph(state, graph, DENOISE_LATENTS);
+  await addControlNetToLinearGraph(state, graph, SDXL_DENOISE_LATENTS);
 
   // Add IP Adapter
-  await addIPAdapterToLinearGraph(state, graph, DENOISE_LATENTS);
-  await addT2IAdaptersToLinearGraph(state, graph, DENOISE_LATENTS);
+  await addIPAdapterToLinearGraph(state, graph, SDXL_DENOISE_LATENTS);
+
+  await addT2IAdaptersToLinearGraph(state, graph, SDXL_DENOISE_LATENTS);
+
   // NSFW & watermark - must be last thing added to graph
   if (state.system.shouldUseNSFWChecker) {
     // must add before watermarker!
