+++ invokeai/frontend/web/src/features/nodes/util/graph/generation/Graph.ts
